Create hash_python.py
This file demonstrates the Hash extension attack. 


Here, we try to elevate our privileges by adding ",admins" to the existing permissions.
 We extend the hash by padding the original text to be 8 byte short of 64  and then the length of the string for 8 bytes.
 Now, we append the malicious string we want to add. It takes the input, computes, and returns output in hexadecimal. 

Then it uses the originally found hash as the base values and tries to compute a new hash with the new padding and the altered hash. 

This would match the hash found in the malicious string if it had been added originally.